+++ packages/camera/camera_avfoundation/ios/Classes/FLTThreadSafeEventChannel.m
@@ -21,11 +21,13 @@
 
 - (void)setStreamHandler:(NSObject<FlutterStreamHandler> *)handler
               completion:(void (^)(void))completion {
-  __weak typeof(self) weakSelf = self;
+  // WARNING: Should not use weak self, because FLTThreadSafeEventChannel is a local variable
+  // (retained within call stack, but not in the heap). FLTEnsureToRunOnMainQueue may trigger a
+  // context switch (when calling from background thread), in which case using weak self will always
+  // result in a nil self. Alternative to using strong self, we can also create a local strong
+  // variable to be captured by this block.
   FLTEnsureToRunOnMainQueue(^{
-    typeof(self) strongSelf = weakSelf;
-    if (!strongSelf) return;
-    [strongSelf.channel setStreamHandler:handler];
+    [self.channel setStreamHandler:handler];
     completion();
   });
 }