+++ packages/file_selector/file_selector_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum Flutter version to 2.10.
+
 ## 2.1.0
 
 * Adds `allowsAny` to `XTypeGroup` as a simple and future-proof way of identifying